Add support for rp2350 micro-controllers (#6725) (#455)
* lib: Move lib/rp2040/elf2uf2 to lib/elf2uf2

Recent versions of the rp2040 sdk no longer contain the elf2uf2 tool.
So, move that code to a new dedicated directory.  This is in
preparation for updating the rp2040 sdk version.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* lib: Update lib/rp2040 to v2.0.0 SDK release

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* lib: Update can2040 to support v2.0.0 of pico-sdk

A new version of can2040 is needed due to changes in the 2.0.0 release
of the pico-sdk.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* lib: Add rp2350 files to pico-sdk

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* lib: Add cortex-m33 support files to lib/cmsis-core/

This is in preparation for adding support for rp2350 mcus.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* armcm_boot: Support ARM cortex-m33 chips

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Rename CONFIG_RP2040_yyy Kconfig symbols to CONFIG_RPXXXX_yyy

Rename the Kconfig symbols.  This is in preparation to adding support
for the rp2350 mcu.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Rename rp2040_link.lds.S to rpxxxx_link.lds.S

This is in preparation for rp2350 support.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Avoid using memcpy() on USB dpram

Some versions of the system memcpy() may make unaligned memory
accesses, which can result in a bus fault when accessing the usb dpram
device memory.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Use a higher USB PLL internal frequency

The rp2350 chip requires a higher internal frequency, so choose a
value that works for both rp2040 and rp2350.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Initial rp2350 support

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* lib: Update rp2040_flash to upstream picotool.git v2.0.0

This is in preparation for adding rp2350 flash support.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* lib: Update rp2040_flash code to support rp2350 reboot

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* flash_usb: Initial support for flashing rp2350 chips

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* temperature_mcu: Add support for rp2350 MCUs

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Move chipid reading to bootrom.c

Rewrite chipid.c so that it contains just the USB and canbus id
manipulation code.  Move the low-level chipid reading to bootrom.c.

Also, introduce a new bootrom_reboot_usb_bootloader() function in
bootrom.c so that the main.c code does not need to know the specifics
of rebooting into the bootrom.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Add rp2350 bootrom based chipid and reboot to bootloader code

This adds the bootrom code needed to implement "reboot into
bootloader" and "chipid" capabilities.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* docs: Add rp2350 to benchmarks

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

---------

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>
Co-authored-by: Kevin O'Connor <kevin@koconnor.net>