Improved canbus diagnostics (#6784) (#568)
* stm32: Add support for reporting canbus state from can.c

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* stm32: Add support for reporting canbus state from fdcan.c

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* rp2040: Add support for reporting canbus state

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* atsamd: Add support for reporting canbus state

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* atsam: Add support for reporting canbus state

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* canbus_stats: Periodically report canbus interface statistics

Add support for a new get_canbus_status command to canserial.c .

Add new canbus_stats.py module that will periodically query canbus
mcus for connection status information.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* usb_canbus: Detect canbus stalls when in usb to canbus bridge mode

If the low-level canbus stops working then it could become impossible
to send messages to and from the canbus bridge node itself.  This can
make it difficult to diagnose canbus problems.

Change the canbus bridge code to detect if message transmits become
stalled for 50+ milliseconds and go into a "discarding" state.  In
this discarding state, messages destined for the canbus will be
discarded until the canbus becomes active again.  In this discarding
state it will therefore be possible to transmit messages to and from
the canbus bridge node.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

---------

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>
Co-authored-by: Kevin O'Connor <kevin@koconnor.net>