linux: Use volatile sig_atomic_t for must_wake_timers
The existing code relied on a memory barrier which didn't exist, but
it's non-standard C to do anything but use a volatile sig_atomic_t.
Switch to using a sig_atomic_t for must_wake_timers.

I've also pulled it out of the TimerInfo structure because to me it's
unclear if it's legal to read inside of a non-volatile struct in C.