bed_mesh: reduce generated point logging (#597)
With the introduction of "scanning" probes it has
become common for configurations to generate a large
number of points.  This can overwhelm both the log and
the pty when new points are generated.

This patch limits the initial points logged to 50.  In
addition points are no longer logged or pushed over
the pty when the mesh configuration changes.

Signed-off-by: Eric Callahan <arksine.code@gmail.com>
Co-authored-by: Eric Callahan <arksine.code@gmail.com>