tmc5160: expose driver_CS, defaults to 31 (#570)
* Added driver_cs

The `driver_CS` parameter has been added to tmc5160. The
ideal driver_cs value may be found by setting the CS value on the
tmc5160_calculations.xlsx spreadsheet, under the chopper tab, so that
the Rsense value in the spreadsheet matches `sense_resistor` as defined
in printer.cfg. While it's not necessary to change the CS value, it can
be helpful to reach adequate hystersis values on high current drivers paired
with low current motors. The default for this value is 31, meaning only
globalscaler will be used to scale the current during normal operation.
Errors will be invoked if the CS value is set too low, as the target
current will not be able to be reached.

Update tmc5160.py

* Update tmc5160.py

* Update tmc5160.py

* Update tmc5160.py

* Update tmc5160.py

* Update tmc5160.py

* Changes to Docs

* Update control_mpc.py

* Update control_mpc.py

* docs - improve message

* adds tests

---------

Co-authored-by: Rogerio Goncalves <rogerlz@gmail.com>