usb_cdc: Avoid ending a transmission with a max size usb packet (#599)
* usb_cdc: Avoid ending a transmission with a max size usb packet

It seems the Linux kernel will consider a maximum size usb packet to
be a transaction that will continue into the next usb packet.  It will
thus hold on to the traffic from the first packet until it gets the
next packet.  However, if the mcu has no further data to send after
the first packet then the data could get delayed for an extended
period of time.

To avoid this, check for transmissions that could end on a maximum
sized packet and send that data in two packets instead.  This avoids
this unusual corner case.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* usb_cdc_ep: Define endpoint sizes in usb_cdc_ep.h

Move the definition of the usb endpoint sizes from usb_cdc.h to
usb_cdc_ep.h .  This allows individual boards to override the default
endpoint sizes.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

---------

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>
Co-authored-by: Kevin O'Connor <kevin@koconnor.net>