Add build test for unexpected run-time software divide (#674)
* lcd_hd44780: Make sure nsecs_to_ticks() is always inlined

It is a compile-time calculation that needs to be inlined to work.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* lcd_st7920: Make sure nsecs_to_ticks() is always inlined

It is a compile-time calculation that needs to be inlined to work.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* neopixel: Make sure nsecs_to_ticks() is always inlined

It is a compile-time calculation that needs to be inlined to work.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* Kconfig: Note which chips require software divide operations

Add a new HAVE_SOFTWARE_DIVIDE_REQUIRED that indicates which chips
require software divide.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* check-software-div: Add a new build check for software divide

Update the build checks to include a check for unexpected software
divide operations.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

---------

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>
Co-authored-by: Kevin O'Connor <kevin@koconnor.net>