Support step on both edges with custom minimum pulse duration (#678)
* stepper: Support step on both edges with custom minimum pulse duration

Add support for "step on both edges" to the main stepper_event_full()
code.  This makes that mode of operation available even when the
micro-controller is not compiled for "optimized step on both edges".
It also enables the custom pulse duration support (step_pulse_ticks)
when in "step on both edges" mode.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

* stepper: Support disabling optimized "step on both edges" in "make menuconfig"

Add a new "low level option" to allow users to configure if they want
to optimize for Trinamic drivers or traditional stepper motor drivers.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>

---------

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>
Co-authored-by: Kevin O'Connor <kevin@koconnor.net>