tmctrigorilla: Refactor and merge code in to standard TMC code
The current tmctrigorilla code is copy pasted code that I have to
manually track and update whenever the rest of the TMC code gets
changed. This has not been great to maintain.

Instead of copy pasting code, add a few conditions to the standard TMC
code so it supports acting without readable registers. While we're at
it, support reading cached registers.

Signed-off-by: Jookia <contact@jookia.org>